+++ packages/zowe-explorer/src/trees/job/JobFSProvider.ts
@@ -26,8 +26,10 @@ import {
     FsJobsUtils,
     FsAbstractUtils,
     ZoweExplorerApiType,
+    ZosEncoding,
+    AuthHandler,
 } from "@zowe/zowe-explorer-api";
-import { IJob, IJobFile } from "@zowe/zos-jobs-for-zowe-sdk";
+import { IDownloadSpoolContentParms, IJob, IJobFile } from "@zowe/zos-jobs-for-zowe-sdk";
 import { Profiles } from "../../configuration/Profiles";
 import { ZoweExplorerApiRegister } from "../../extending/ZoweExplorerApiRegister";
 import { SharedContext } from "../shared/SharedContext";
@@ -35,11 +37,14 @@ import { AuthUtils } from "../../utils/AuthUtils";
 
 export class JobFSProvider extends BaseProvider implements vscode.FileSystemProvider {
     private static _instance: JobFSProvider;
+
     private constructor() {
         super();
         ZoweExplorerApiRegister.addFileSystemEvent(ZoweScheme.Jobs, this.onDidChangeFile);
         this.root = new DirEntry("");
     }
+
+    public encodingMap: Record<string, ZosEncoding> = {};
 
     public static get instance(): JobFSProvider {
         if (!JobFSProvider._instance) {
@@ -203,20 +208,31 @@ export class JobFSProvider extends BaseProvider implements vscode.FileSystemProv
         const bufBuilder = new BufferBuilder();
 
         const jesApi = ZoweExplorerApiRegister.getJesApi(spoolEntry.metadata.profile);
+        await AuthHandler.lockProfile(spoolEntry.metadata.profile);
         try {
             if (jesApi.downloadSingleSpool) {
-                await jesApi.downloadSingleSpool({
+                const spoolDownloadObject: IDownloadSpoolContentParms = {
                     jobFile: spoolEntry.spool,
                     stream: bufBuilder,
-                });
+                };
+
+                // Handle encoding and binary options
+                if (spoolEntry.encoding) {
+                    spoolDownloadObject.binary = spoolEntry.encoding.kind === "binary";
+                    if (spoolEntry.encoding.kind === "other") {
+                        spoolDownloadObject.encoding = spoolEntry.encoding.codepage;
+                    }
+                }
+                await jesApi.downloadSingleSpool(spoolDownloadObject);
             } else {
                 const jobEntry = this._lookupParentDirectory(uri) as JobEntry;
                 bufBuilder.write(await jesApi.getSpoolContentById(jobEntry.job.jobname, jobEntry.job.jobid, spoolEntry.spool.id));
             }
         } catch (err) {
-            AuthUtils.promptForAuthError(err, spoolEntry.metadata.profile);
+            await AuthUtils.handleProfileAuthOnError(err, spoolEntry.metadata.profile);
             throw err;
         }
+        AuthHandler.unlockProfile(spoolEntry.metadata.profile);
 
         this._fireSoon({ type: vscode.FileChangeType.Changed, uri });
         spoolEntry.data = bufBuilder.read() ?? new Uint8Array();